services/rag/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.4
pydantic==2.6.1
duckdb==0.10.0
